Permission PrioriteController*

+++ src/Controller/PrioriteController.php
@@ -5,7 +5,9 @@ namespace App\Controller;
 use App\Entity\Priorite;
 use App\Form\PrioriteType;
 use App\Repository\PrioriteRepository;
+use App\Services\Personne\PermissionChecker;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
@@ -16,13 +18,36 @@ use Symfony\Component\Routing\Annotation\Route;
 class PrioriteController extends AbstractController
 {
     /**
+     * @var PermissionChecker
+     */
+    private $permissionChecker;
+
+    /**
+     * PrioriteController constructor.
+     */
+    public function __construct(PermissionChecker $permissionChecker)
+    {
+        $this->permissionChecker = $permissionChecker;
+    }
+
+    /**
      * @Route("/", name="priorite_index", methods={"GET"})
      */
     public function index(PrioriteRepository $prioriteRepository): Response
     {
-        return $this->render('priorite/index.html.twig', [
-            'priorites' => $prioriteRepository->findAll(),
-        ]);
+        if(!$this->isGranted('ROLE_USER')){
+            $this->addFlash('fail','Veuillez vous connectez pour acceder à cette page.');
+            return $this->redirectToRoute('app_login');
+        }else {
+            if (!$this->permissionChecker->isUserGranted(["GET_OTHER_PRIORITE"])) {
+                $this->addFlash('fail', 'Vous ne possedez pas les permissions necessaires.');
+                return new RedirectResponse("/");
+            } else {
+                return $this->render('priorite/index.html.twig', [
+                    'priorites' => $prioriteRepository->findAll(),
+                ]);
+            }
+        }
     }
 
     /**
@@ -30,22 +55,32 @@ class PrioriteController extends AbstractController
      */
     public function new(Request $request): Response
     {
-        $priorite = new Priorite();
-        $form = $this->createForm(PrioriteType::class, $priorite);
-        $form->handleRequest($request);
+        if(!$this->isGranted('ROLE_USER')){
+            $this->addFlash('fail','Veuillez vous connectez pour acceder à cette page.');
+            return $this->redirectToRoute('app_login');
+        }else {
+            if (!$this->permissionChecker->isUserGranted(["POST_PRIORITE"])) {
+                $this->addFlash('fail', 'Vous ne possedez pas les permissions necessaires.');
+                return new RedirectResponse("/");
+            } else {
+                $priorite = new Priorite();
+                $form = $this->createForm(PrioriteType::class, $priorite);
+                $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->persist($priorite);
-            $entityManager->flush();
+                if ($form->isSubmitted() && $form->isValid()) {
+                    $entityManager = $this->getDoctrine()->getManager();
+                    $entityManager->persist($priorite);
+                    $entityManager->flush();
 
-            return $this->redirectToRoute('priorite_index');
+                    return $this->redirectToRoute('priorite_index');
+                }
+
+                return $this->render('priorite/new.html.twig', [
+                    'priorite' => $priorite,
+                    'form' => $form->createView(),
+                ]);
+            }
         }
-
-        return $this->render('priorite/new.html.twig', [
-            'priorite' => $priorite,
-            'form' => $form->createView(),
-        ]);
     }
 
     /**
@@ -53,9 +88,19 @@ class PrioriteController extends AbstractController
      */
     public function show(Priorite $priorite): Response
     {
-        return $this->render('priorite/show.html.twig', [
-            'priorite' => $priorite,
-        ]);
+        if(!$this->isGranted('ROLE_USER')){
+            $this->addFlash('fail','Veuillez vous connectez pour acceder à cette page.');
+            return $this->redirectToRoute('app_login');
+        }else {
+            if (!$this->permissionChecker->isUserGranted(["GET_OTHER_PRIORITE"])) {
+                $this->addFlash('fail', 'Vous ne possedez pas les permissions necessaires.');
+                return new RedirectResponse("/");
+            } else {
+                return $this->render('priorite/show.html.twig', [
+                    'priorite' => $priorite,
+                ]);
+            }
+        }
     }
 
     /**
@@ -63,19 +108,29 @@ class PrioriteController extends AbstractController
      */
     public function edit(Request $request, Priorite $priorite): Response
     {
-        $form = $this->createForm(PrioriteType::class, $priorite);
-        $form->handleRequest($request);
+        if(!$this->isGranted('ROLE_USER')){
+            $this->addFlash('fail','Veuillez vous connectez pour acceder à cette page.');
+            return $this->redirectToRoute('app_login');
+        }else {
+            if (!$this->permissionChecker->isUserGranted(["UPDATE_OTHER_PRIORITE"])) {
+                $this->addFlash('fail', 'Vous ne possedez pas les permissions necessaires.');
+                return new RedirectResponse("/");
+            } else {
+                $form = $this->createForm(PrioriteType::class, $priorite);
+                $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
+                if ($form->isSubmitted() && $form->isValid()) {
+                    $this->getDoctrine()->getManager()->flush();
 
-            return $this->redirectToRoute('priorite_index');
+                    return $this->redirectToRoute('priorite_index');
+                }
+
+                return $this->render('priorite/edit.html.twig', [
+                    'priorite' => $priorite,
+                    'form' => $form->createView(),
+                ]);
+            }
         }
-
-        return $this->render('priorite/edit.html.twig', [
-            'priorite' => $priorite,
-            'form' => $form->createView(),
-        ]);
     }
 
     /**
@@ -83,10 +138,20 @@ class PrioriteController extends AbstractController
      */
     public function delete(Request $request, Priorite $priorite): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$priorite->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($priorite);
-            $entityManager->flush();
+        if(!$this->isGranted('ROLE_USER')){
+            $this->addFlash('fail','Veuillez vous connectez pour acceder à cette page.');
+            return $this->redirectToRoute('app_login');
+        }else {
+            if (!$this->permissionChecker->isUserGranted(["DELETE_OTHER_PRIORITE"])) {
+                $this->addFlash('fail', 'Vous ne possedez pas les permissions necessaires.');
+                return new RedirectResponse("/");
+            } else {
+                if ($this->isCsrfTokenValid('delete' . $priorite->getId(), $request->request->get('_token'))) {
+                    $entityManager = $this->getDoctrine()->getManager();
+                    $entityManager->remove($priorite);
+                    $entityManager->flush();
+                }
+            }
         }
 
         return $this->redirectToRoute('priorite_index');
